requirements.txt
flask
python-dotenv
pandas
matplotlib
openpyxl
